feat : 레포지토리 작성 및 레포지토리 단위 테스트 완료
UserJPARepository, BoardJPARepository 작성 및 단위테스트 완료

+++ src/main/java/wanted/onboarding/board/Board.java
@@ -15,7 +15,7 @@ import javax.persistence.*;
 public class Board {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     @ManyToOne(fetch = FetchType.LAZY)
     private User user;
     @Column
@@ -24,7 +24,7 @@ public class Board {
     private String content;
 
     @Builder
-    public Board(int id, User user, String title, String content) {
+    public Board(Long id, User user, String title, String content) {
         this.id = id;
         this.user = user;
         this.title = title;
